+++ Project2/minix/commands/partition/partition.c
@@ -13,7 +13,6 @@
 #include <sys/ioctl.h>
 #include <limits.h>
 #include <stdint.h>
-#include <inttypes.h>
 #include <assert.h>
 
 #ifdef __minix
@@ -192,8 +191,7 @@ void show_part(struct part_entry *p)
 	printf("%3d ", (n-1) / 2);
 	show_chs(p->lowsec);
 	show_chs(p->lowsec + p->size - 1);
-	printf("  %8"PRIu32"  %8"PRIu32"  %7"PRIu32"\n",
-	    p->lowsec, p->size, p->size / 2);
+	printf("  %8lu  %8lu  %7lu\n", p->lowsec, p->size, p->size / 2);
 }
 
 void usage(void)
@@ -375,8 +373,8 @@ void distribute(void)
 			if (pe->bootind & EXIST_FLAG) {
 				if (base > pe->lowsec) {
 					fprintf(stderr,
-	"%s: fixed partition %u is preceded by too big partitions/holes\n",
-				arg0, ((unsigned int)(pe - table) - 1) / 2);
+	"%s: fixed partition %ld is preceded by too big partitions/holes\n",
+						arg0, ((pe - table) - 1) / 2);
 					exit(1);
 				}
 				exp= nil;	/* XXX - Extend before? */
